requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.2
altair==5.5.0
altex==0.2.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
backports.tarfile==1.2.0
beartype==0.22.5
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.41.5
botocore==1.41.5
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.2.4
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
duckduckgo_search==8.1.1
email-validator==2.3.0
emoji==2.15.0
entrypoints==0.4
et_xmlfile==2.0.0
Events==0.5
exceptiongroup==1.3.0
faiss-cpu==1.13.0
Faker==38.2.0
fastapi==0.121.3
fastmcp==2.13.1
favicon==0.7.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htbuilder==0.9.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.6
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
llvmlite==0.45.1
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
markdownify==1.2.2
markdownlit==0.0.7
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mcp==1.21.2
mdurl==0.1.2
mem0ai==1.0.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
olefile==0.47
ollama==0.6.1
openai==2.8.1
openapi-pydantic==0.5.1
openpyxl==3.1.5
opensearch-py==3.0.0
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
pdf2image==1.17.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
portalocker==3.2.0
posthog==7.0.1
primp==0.15.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.17.2
pynndescent==0.5.13
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdf==6.3.0
PyPDF2==3.0.1
pyperclip==1.11.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qdrant-client==1.16.1
RapidFuzz==3.14.3
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
s3transfer==0.15.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.1.0
snowflake-snowpark-python==1.42.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.50.0
streamlit==1.51.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-shadcn-ui==0.1.19
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
unstructured==0.18.20
unstructured-client==0.42.4
urllib3==2.5.0
uvicorn==0.38.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zep-python==2.0.2
zipp==3.23.0
zstandard==0.25.0
